events-v3/jwt_microservice/requirements.txt
# This file lists the Python packages required to run the FastAPI application.
# pip installs these dependencies when building the Docker image.
fastapi
uvicorn
PyJWT
python-dotenv
